Add files via upload
This code is for creating the notes that you will upload to the Arduino code. With this website I created, you will be able to play most of one-handed songs without having to manually type every single key to the Arduino code. The index is the HTML file and sketch is the javaScript that holds the code.